+++ docs/change_log.md
@@ -73,6 +73,7 @@ There is one significant exception to the above rules -- `MAJOR=0` releases. Any
 
 - renamed the `simmate engine` command group to `simmate compute` and updated internal modules/tests
 - refactored `website.data_explorer` to use HTMX components for all table entries, moving UI-specific features from `DatabaseTable` and `HTMLMixin` into `DynamicTableForm`
+- moved `chemspace.utils.download_raw_files` into `ChemSpaceFreedomSpaceMolecule.download_source_data`
 
 **Fixes**